fix: resolve four endpoint mapping and tab bugs
- parse __body__ string back to JSON after Jinja render so it is sent
  as a JSON object rather than a quoted string literal
- re-add $. prefix when rebuilding resBody from parseResMapping so
  backend treats values as JSONPath expressions
- initialise and sync TestAndMap pathToVar from responseMapping prop
  so existing mappings pre-populate on load and after auto-configure
- wire EndpointHeadersTab at index 2 in EndpointDetailTabs; remove
  orphaned EndpointMappingsTab that was incorrectly occupying that slot
